average

+++ 1208_2.c
@@ -4,7 +4,7 @@
 
 double solution(int arr[], size_t arr_len) {
     double answer = 0;
-    for (int i=0; i<arr_len; i++)
+    for (size_t i=0; i<arr_len; i++)
         answer += arr[i];
     answer /= arr_len;
     return answer;
